+++ NEWS.md
@@ -1,6 +1,11 @@
 # mlr3tuning (development version)
 
+# mlr3tuning 0.19.0
+
+* fix: Add `$phash()` method to `AutoTuner`.
+* fix: Include `Tuner` in hash of  `AutoTuner`.
 * feat: Add new callback that scores the configurations on additional measures while tuning.
+* feat: Add vignette about adding new tuners which was previously part of the mlr3book.
 
 # mlr3tuning 0.18.0